requirements.txt
dataclass-wizard
GitPython
more_itertools
python-doctr
torch
tqdm
